Report unused ignore lines in config
This change allows traceroute to report when there are unused lines in its config. ie, lines that never "match to ignore" anything.

Unused lines are not just a buildup of historic cruft, they could also unintentionally ignore matching code further down the line if the regex is left in, and then it matches something else entirely.